requirements.txt
crewai==0.5.3
pandas
email
openai
python-dotenv